ovsdb: Rename ovsdb_file_replica to ovsdb_file.
This is in preparation for exposing ovsdb_file to clients outside this
translation unit.  These clients don't care that the ovsdb_file is an
ovsdb replica--that's an implementation detail--and so it makes sense to
rename it from their point of view.

This is just a search-and-replace plus reindenting where appropriate.