ofproto: Start work to enable datapaths with built-in wildcard support.
The "xflow" interface assumes that the datapath supports only exact-match
flows.  This seems to be one reasonable way to implement a datapath in
software, but some hardware datapaths have built-in support for flows that
include wildcards.  Until now, OVS has not been able to take advantage of
this feature.  This commit is a beginning of that support.  It works, to
some extent, but it is incomplete in many ways, and should be considered
a prototype.