netdev: New function netdev_reopen().
In the new wdp layer it is convenient to pass around copies of struct
wdp_port, which contains a struct netdev *.  To make this efficient
it makes sense to maintain a reference count in struct netdev itself, so
that copying and freeing copies of a netdev become just incrementing and
decrementing the reference count.  This commit adds that ability.