Fix EXTRA_DIST dependence on Automake conditionals.
Automake respects conditionals around EXTRA_DIST assignments.  That is, if
COND is not true, then the following will not distribute 'myfile':
    if COND
    EXTRA_DIST += myfile
    endif

See http://article.gmane.org/gmane.comp.sysutils.automake.general/10891
for more information.

This behavior is surprising, at least to me.  But we can work around it:
anything that can ever *potentially* be assigned to noinst_HEADERS is
always distributed.  So this commit eliminates the problem by adding
$(EXTRA_DIST) to noinst_HEADERS.