xenserver: Ditch ovs-xenserverd even on XenServer 5.5.0.
Examining the XAPI source code shows that at startup it invokes a script
named /opt/xensource/libexec/xapi-startup-script, if one exists.  Testing
shows that this was also true in XenServer 5.5.0.  No such script is
installed by default.  Searching for "xapi-startup-script" on Google (with
the quotes) has only one hit, which is documentation on XAPI startup.  So
it seems that we're pretty safe in taking advantage of this hook ourselves.

This commit changes the RPM scripts to install refresh-network-uuids as
the XAPI startup hook on XenServer 5.5.0.

CC: Rob Hoes <rob.hoes@citrix.com>