xenserver: Don't start OVS if network.conf isn't "openvswitch"
The OVS processes would start as long as "/etc/xensource/network.conf"
didn't contain "bridge".  The other OVS scripts, however, would complain
(and not run) if it wasn't "vswitch" or "openvswitch".  This commit will
only start the OVS processes if a value of "vswitch" or "openvswitch" is
present, so some consistency is provided.

(If "/etc/xensource/network.conf" is not "bridge", "vswitch", or
"openvswitch", then XAPI will refuse to run anyway, so not much is going
to happen on the system.)