ovsdb-tool: Don't wait for lockfiles on "compact" or "convert".
Until now, "ovsdb-tool convert" and "ovsdb-tool compact" were willing to
wait an arbitrarily long time to get a lock on the database or on a
temporary file.  In practice, though, the program most likely to have the
lock is ovsdb-server, which is not a transient process, and thus
ovsdb-tool will likely wait indefinitely if it waits at all.

This commit changes these ovsdb-tool commands so that they won't wait at
all if some other program has the lock.  (The other commands already had
this behavior.)

As a side effect, this change fixes the behavior of
"/etc/init.d/openvswitch start" when Open vSwitch was already running.
Previously, it would hang waiting for "ovsdb-tool convert" to finish.
Now, it completes quickly with an error.

CC: Reid Price <reid@nicira.com>
Bug #2864.