bridge: Remove commented-out code for custom management and snoop listeners.
Before the transition from configuration file to OVSDB, it was possible to
override the defaults for OpenFlow management listeners and for OpenFlow
controller connection snooping.  The former can now effectively be done by
configuring a controller through the database.  Overriding the latter is
not very useful (no one has complained that it cannot be done any longer).
So this commit deletes the commented-out code.