ovsdb-idlc: Check and restore ovsdb_datum invariants when setting data.
ovsdb_datum has an important invariant (documented in the large comment
on the declaration of struct ovsdb_datum) that a lot of code relies upon:
keys must be unique and in sorted order.  Most code that creates or
modifies ovsdb_datum structures maintains this invariant.  However, the
"set" functions generated by ovsdb-idlc.in do not check or restore it.
This commit adds that checking.

This might fix an actual bug or two--none have been reported--but mostly it
is just to add some additional checking to avoid future subtle bugs.