vswitch: Distinguish mirrors by UUID, not by name.
A "feature" that ovs-vswitchd inherited from its previous form of
configuration is that every mirror has a name.  Names are not necessarily
meaningful, and there is no reason that they should be unique.  But the
existing implementation depends on them being unique within a given
bridge, and if they are not drops one of the duplicates.

This commit drops the uniqueness requirement.  Instead, it distinguishes
mirrors based on UUID alone.

This commit does not drop the concept of names for mirrors.  There is no
technical reason to retain them, but it is possible that users find them
useful for management reasons.  The names appear in log messages related
to mirrors, which may make the messages easier to understand.

Bug #2416.