datapath: Don't compute checksums on partial packets.
If we are only copying part of a packet to userspace don't bother
computing the checksum on that part since it is meaningless.
Instead, fall back to the old method of checksumming before
copying to ensure the correct result.

This was supposed to be part of the previous commit but was left
off.