tests: Fix occasional failure of "self-linking idl, inconsistent ops" test.
Commit 97f7803 made this test nondeterministic: depending on the ordering
of the random row UUIDs, the error message might mention one of two
different nonexistent rows that are referenced.  This commit fixes the
problem by accepting either one of the two correct rows.

Previous related fixes took a different approach, of instead making the
test deterministic.  But now I think that that is not as good an idea,
because we need to be able to test that OVSDB works whether the query is
deterministic or not.

(Of course, every OVSDB test is nondeterministic, since the UUIDs are
random.  This test was just more nondeterministic than most.)