timeval: Hide "memory leak" from Valgrind.
In glibc, "timer_t" is a "void *" that appears to point into malloc()'d
memory.  By throwing it away entirely, we leak it, which makes valgrind
complain.  We really don't ever care to use the timer object again, but
we can't destroy it without stopping the periodic timer.  So make it
static to avoid a warning from Valgrind.