ovs-pki: Allow generating certificates with duplicate subjects.
Without this setting, the certificate authorities that ovs-pki creates will
not allow two switches or two controllers to have the same name.  This
causes problem in testing, since it's often convenient to test with short,
common names like "tmp".

(If you need to fix a PKI that you already created, in addition to
modifying ca.cnf you will need to make the same change to index.txt.attr.)

CC: Pierre Ettori <pettori@nicira.com>