wdp: Add new wdp_class member function 'get_table_stats'.
The wdp_stats structure is really a relic from the "master" branch.  Many
of its members do not make sense in a switch that might have an arbitrary
number of tables.  There are only a few users of this structure.  This
commit gets rid of one of these users by adding a new member function in
place of building generic functionality on top of wdp_stats information.