ovsdb: Extend "monitor" to select different operations in a single table.
Until now, "monitor" has only allowed the client to choose the kinds of
changes that will be monitored on a per-table basis.  However, it makes
sense to be able to choose operations on a per-column basis.  The
immediate need for this is to make sure that the final statistics of
deleted Interface records are known at time of deletion, even though the
intermediate values of the statistics are not important.

CC: Jeremy Stribling <strib@nicira.com>