datapath: Add loop checking.
New types of vports such as patch and GRE make it possible to
connect multiple (or the same) datapaths together, which introduces
the possibility of loops on a single machine.  Local loops are
even worse than network loops because they will quickly crash the
machine once the kernel stack is exhausted.  This adds loop
checking within the OVS datapath that will drop packets that have
been seen more than 5 times.

CC: Paul Ingram <paul@nicira.com>