classifier: Fix classifier bugs.
The classifier tests have been failing on the 'wdp' branch for a long time
now, ever since commit 8980c78 "ofproto: Start work to enable datapaths
with built-in wildcard support."  I finally got around to finding the
cause, which is that the classifier was hashing and comparing the
'priority' member of exact-match flows.  That member is basically
meaningless for such flows, so this commit fixes the problem by changing
the functions that hash and compare flows to ignore it.  They also now
ignore the 'wildcards' field.  These functions' callers already assumed
these semantics, so there is no change to the classifier itself.  The
classifier tests, on the other hand, were not treating the priority of
exact-match flows in the same way as the classifier itself, so this commit
changes the tests' behavior to match.