Tests: Fix nonportable \" escape in printf(1) invocation.
The \" escape is not part of POSIX, but it is a common extension.  The
dash shell's built-in "printf" implementation does not include this
extension, which caused the testsuite to be generated incorrectly if it
is used as the default shell (as it is on newer versions of Debian and
Ubuntu).

However, there's no reason to use a backslash here at all, so just omit
it.