doc: Make explicit that ovs-vswitchd is the preferred switch.
Many people who are looking for an OpenFlow switch assume that
the only way to do it is using ovs-openflowd.  Sometimes they also
run ovs-vswitchd at the same time.  Try to clarify that neither
of these are necessary and ovs-vswitchd can handle OpenFlow by
itself and is the preferred method of doing so.